refactor beacon node details to tabs
