+++ src/renderer/containers/BeaconNode/dashboard/BeaconNodeDashboard.tsx
@@ -1,27 +1,19 @@
 import React, {useEffect, useState} from "react";
-import {LogStream} from "../../components/LogStream/LogStream";
-import {BeaconChain} from "../../services/docker/chain";
-import {DockerRegistry} from "../../services/docker/docker-registry";
-import {Beacon, BeaconStatus} from "../../ducks/beacon/slice";
+import {Beacon, BeaconStatus} from "../../../ducks/beacon/slice";
 import {BeaconNodeResponseTimeChart} from "./BeaconNodeResponseTimeChart";
 import {BeaconNodeResponseErrorPieChart, ResponseErrorPieData} from "./BeaconNodeResponseErrorPieChart";
-import database from "../../services/db/api/database";
-import {SimpleLineChartRecord} from "../../components/SimpleLineChart/SimpleLineChart";
-import {getLatencyChartData, getNetworkErrorPieData} from "../../services/utils/charts";
+import database from "../../../services/db/api/database";
+import {SimpleLineChartRecord} from "../../../components/SimpleLineChart/SimpleLineChart";
+import {getLatencyChartData, getNetworkErrorPieData} from "../../../services/utils/charts";
 import ReactTooltip from "react-tooltip";
-import {capitalize} from "../../services/utils/formatting";
+import {capitalize} from "../../../services/utils/formatting";
+import {BeaconNodeButtons} from "../../BeaconNodes/BeaconNodeButtons";
 
 interface IBeaconNodeProps {
     beacon: Beacon;
-    showTitle?: boolean;
 }
 
-export const BeaconNode: React.FC<IBeaconNodeProps> = ({
-    beacon: {url, network, slot, status, docker},
-    showTitle = true,
-}) => {
-    const container = docker && (DockerRegistry.getContainer(docker.id) as BeaconChain);
-
+export const BeaconNodeDashboard: React.FC<IBeaconNodeProps> = ({beacon: {url, network, slot, status, docker}}) => {
     const [avgLatency, setAvgLatency] = useState<SimpleLineChartRecord[]>([]);
     const [avgLatencyTicks, setAvgLatencyTicks] = useState<string[]>([]);
     const [pieData, setPieData] = useState<ResponseErrorPieData>([
@@ -51,12 +43,10 @@ export const BeaconNode: React.FC<IBeaconNodeProps> = ({
 
     return (
         <div className='flex-column stretch'>
-            {showTitle && (
-                <div className='row space-between' style={{paddingTop: "25px"}}>
-                    <h2>Beacon Node</h2>
-                    <h5>{url}</h5>
-                </div>
-            )}
+            <div className='row space-between' style={{paddingTop: "25px"}}>
+                <h2>Beacon Node</h2>
+                <h5>{url}</h5>
+            </div>
 
             <div className='row space-between'>
                 <h2>{capitalize(network)}</h2>
@@ -89,13 +79,9 @@ export const BeaconNode: React.FC<IBeaconNodeProps> = ({
                 <BeaconNodeResponseTimeChart data={avgLatency} ticks={avgLatencyTicks} />
                 <BeaconNodeResponseErrorPieChart data={pieData} />
             </div>
+            <br />
 
-            {container && (
-                <div className='box log-stream-container'>
-                    <h4>Log Stream</h4>
-                    <LogStream source={container ? container.getLogs()! : undefined} />
-                </div>
-            )}
+            {docker && <BeaconNodeButtons image={docker.id} url={url} />}
         </div>
     );
 };
