+++ src/renderer/containers/BeaconNode/dashboard/BeaconNodeResponseErrorPieChart.tsx
